♻️ vue-dot: Refactor warning rules into mixin
